+++ transport/rest/handler/types.go
@@ -2,7 +2,7 @@ package restHandler
 
 import (
 	"github.com/alhamsya/boilerplate-go/lib/helpers/config"
-	"github.com/alhamsya/boilerplate-go/transport/inter/rest/routers"
+	"github.com/alhamsya/boilerplate-go/transport/rest/routers"
 	"github.com/gofiber/fiber/v2"
 )
